Exception Handling in Service and Controller Layers

+++ gc-backend/src/main/java/com/gameclub/service/MemeService.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Service;
 
 import com.gameclub.dao.MemeDao;
 import com.gameclub.dao.UserDao;
+import com.gameclub.exceptions.MemeNotFoundException;
+import com.gameclub.exceptions.UserNotFoundException;
 import com.gameclub.model.Meme;
 import com.gameclub.model.User;
 
@@ -24,13 +26,17 @@ public class MemeService {
 		return m;
 	}
 	
-	public Meme getMemeById(int id) {
+	public Meme getMemeById(int id) throws MemeNotFoundException {
 		Meme m = mDao.getMemeById(id);
+		if (m == null)
+			throw new MemeNotFoundException();
 		return m;
 	}
 	
-	public Meme getMemeByTitle(String title) {
+	public Meme getMemeByTitle(String title) throws MemeNotFoundException {
 		Meme m = mDao.getMemeByTitle(title);
+		if (m == null)
+			throw new MemeNotFoundException();
 		return m;
 	}
 	
@@ -39,11 +45,15 @@ public class MemeService {
 		return memes;
 	}
 	
-	public void removeMeme(int id) {
+	public void removeMeme(int id) throws MemeNotFoundException {
+		if (!mDao.existsById(id))
+			throw new MemeNotFoundException();
 		mDao.deleteById(id);
 	}
 	
-	public List<Meme> getMemesByAuthor(int userId) {
+	public List<Meme> getMemesByAuthor(int userId) throws UserNotFoundException {
+		if (!uDao.existsById(userId))
+			throw new UserNotFoundException();
 		User u = uDao.getUserById(userId);
 		List<Meme> memes = mDao.getMemesByAuthor(u);
 		return memes;
