+++ gc-backend/src/main/java/com/gameclub/service/UserService.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.gameclub.dao.UserDao;
+import com.gameclub.exceptions.UserAlreadyExistsException;
+import com.gameclub.exceptions.UserNotFoundException;
 import com.gameclub.model.User;
 
 import lombok.AllArgsConstructor;
@@ -16,18 +18,24 @@ public class UserService {
 	
 	private UserDao uDao;
 	
-	public User registerUser(User user) {
+	public User registerUser(User user) throws UserAlreadyExistsException {
+		if (uDao.existsById(user.getId()))
+			throw new UserAlreadyExistsException();
 		User u = uDao.save(user);
 		return u;
 	}
 	
-	public User getUserById(int id) {
+	public User getUserById(int id) throws UserNotFoundException {
 		User u = uDao.getUserById(id);
+		if (u == null)
+			throw new UserNotFoundException();
 		return u;
 	}
 	
-	public User getUserByUsername(String username) {
+	public User getUserByUsername(String username) throws UserNotFoundException {
 		User u = uDao.getUserByUsername(username);
+		if (u == null)
+			throw new UserNotFoundException();
 		return u;
 	}
 	
@@ -36,7 +44,9 @@ public class UserService {
 		return users;
 	}
 	
-	public void removeUser(int id) {
+	public void removeUser(int id) throws UserNotFoundException {
+		if (!uDao.existsById(id))
+			throw new UserNotFoundException();
 		uDao.deleteById(id);
 	}
